requirements.txt
Django>=2.1
python-telegram-bot>=11.0.0